Split TerminalTestReporter.Summary.cs into focused partial files
TerminalTestReporter.Summary.cs had grown to 715 lines and mixed several
unrelated concerns. Move each into its own partial file, following the
existing TerminalTestReporter.<Concern>.cs convention:

- TerminalTestReporter.Coverage.cs: AppendCoverageSummary (public and
  private), FormatThresholdPercentages and the coverage label helpers.
- TerminalTestReporter.TestDiscovery.cs: both TestDiscovered overloads
  and AppendTestDiscoverySummary. Named to pair with the existing
  TerminalTestReporter.TestCompletion.cs, its mirror concern.
- TerminalTestReporter.FlakyTests.cs: AppendFlakyTests and
  AppendFlakyTestLine, which render the optional "Flaky tests" section.
- TerminalTestReporter.SlowestTests.cs: AppendSlowestTests and
  AppendSlowestTestLine.

TerminalTestReporter.Summary.cs keeps AppendTestRunSummary and
AppendRetrySummaryLines; the latter emits the "flaky:" / "retried:"
counter lines inline between the skipped count and the duration, so it
belongs with the rest of the counter block rather than with the separate
flaky-test listing. Every file is now well under 350 lines.

No member body was changed: the 606 non-empty class-body lines of the
original file and of the five files that replace it are identical. Two
private helpers did change position relative to the whole file, each
moving next to its sole caller: AppendSlowestTestLine (previously after
AppendFlakyTestLine) and AppendRetrySummaryLines. The other non-move
change is narrowing each file's using set to what it actually needs.

Also add a <remarks> map of the 13 partials to the type doc on
TerminalTestReporter.cs, following the precedent in
TrxResultStreamingStore.cs.

Fixes #10367

Copilot-Session: d3b865a0-de90-4543-a493-9538aa8f1e2a
