+++ src/Platform/Microsoft.Testing.Platform/OutputDevice/Terminal/TerminalTestReporter.Summary.cs
@@ -1,10 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Testing.Platform.Extensions.Messages;
-using Microsoft.Testing.Platform.Helpers;
-using Microsoft.Testing.Platform.Services;
-
 namespace Microsoft.Testing.Platform.OutputDevice.Terminal;
 
 [UnsupportedOSPlatform("browser")]
@@ -225,193 +221,6 @@ internal sealed partial class TerminalTestReporter
     }
 
     /// <summary>
-    /// Orchestrator overload (<c>dotnet test</c>): the multi-process orchestrator also knows each discovered test's
-    /// uid, file path and line number. The shared discovery summary currently lists display names only, so those are
-    /// accepted for signature parity. When <paramref name="displayName"/> is missing the <paramref name="uid"/> is used
-    /// as the listed name; when neither is available the test is still counted (so the discovery total stays correct)
-    /// but no blank entry is added to the summary.
-    /// </summary>
-    internal void TestDiscovered(string executionId, string? displayName, string? uid, string? filePath, int? lineNumber)
-    {
-        // Prefer the display name, fall back to the uid so the discovered test is still listed by something.
-        string? name = displayName ?? uid;
-        if (name is not null)
-        {
-            TestDiscovered(executionId, name);
-            return;
-        }
-
-        // No name available at all: still increment the discovered count so the discovery summary total stays
-        // correct (in discovery mode TotalTests is computed from DiscoveredTests), but avoid adding a blank entry.
-        if (!_assemblies.TryGetValue(executionId, out TestProgressState? asm))
-        {
-            throw ApplicationStateGuard.Unreachable();
-        }
-
-        asm.ReportDiscoveredTest(displayName: null);
-        _terminalWithProgress.UpdateWorker(asm.SlotIndex);
-    }
-
-    internal void TestDiscovered(string executionId, string displayName)
-    {
-        if (!_assemblies.TryGetValue(executionId, out TestProgressState? asm))
-        {
-            throw ApplicationStateGuard.Unreachable();
-        }
-
-        // In discovery mode TotalTests is computed from DiscoveredTests; in execution mode it is computed from the
-        // passed/skipped/failed tally as tests complete. So we only need to bump the discovered count here.
-        asm.ReportDiscoveredTest(MakeControlCharactersVisible(displayName, true));
-
-        _terminalWithProgress.UpdateWorker(asm.SlotIndex);
-    }
-
-    public void AppendTestDiscoverySummary(ITerminal terminal)
-    {
-        List<TestProgressState> assemblies = [.. _assemblies.Values.OrderBy(static a => a.Id)];
-        terminal.AppendLine();
-
-        int totalTests = assemblies.Sum(static a => a.TotalTests);
-        bool runFailed = WasCancelled || totalTests < 1;
-
-        if (_options.ShowAssembly)
-        {
-            // Orchestrator (dotnet test): a per-assembly "Discovered N tests in assembly - <link>" header followed by
-            // the discovered test names, then a run-level total ("Discovered N tests." / "... in N assemblies.").
-            foreach (TestProgressState assembly in assemblies)
-            {
-                terminal.Append(string.Format(CultureInfo.CurrentCulture, TerminalResources.DiscoveredTestsInAssembly, assembly.DiscoveredTests));
-                terminal.Append(" - ");
-                AppendAssemblyLinkTargetFrameworkAndArchitecture(terminal, assembly);
-                terminal.AppendLine();
-                foreach (string displayName in assembly.DiscoveredTestDisplayNames)
-                {
-                    terminal.Append(SingleIndentation);
-                    terminal.AppendLine(displayName);
-                }
-
-                terminal.AppendLine();
-            }
-
-            terminal.SetColor(runFailed ? TerminalColor.DarkRed : TerminalColor.DarkGreen);
-            terminal.AppendLine(assemblies.Count <= 1
-                ? string.Format(CultureInfo.CurrentCulture, TerminalResources.DiscoveredTestsSummarySingular, totalTests)
-                : string.Format(CultureInfo.CurrentCulture, TerminalResources.DiscoveredTestsSummary, totalTests, assemblies.Count));
-            terminal.ResetColor();
-            terminal.AppendLine();
-
-            if (WasCancelled)
-            {
-                terminal.Append(TerminalResources.Aborted);
-                terminal.AppendLine();
-            }
-
-            return;
-        }
-
-        // In-process host: the single "Test discovery summary: found N test(s)" format (unchanged shipping output).
-        foreach (TestProgressState assembly in assemblies)
-        {
-            foreach (string displayName in assembly.DiscoveredTestDisplayNames)
-            {
-                terminal.Append(SingleIndentation);
-                terminal.AppendLine(displayName);
-            }
-        }
-
-        terminal.AppendLine();
-
-        terminal.SetColor(runFailed ? TerminalColor.DarkRed : TerminalColor.DarkGreen);
-        terminal.Append(string.Format(CultureInfo.CurrentCulture, TerminalResources.TestDiscoverySummarySingular, totalTests));
-
-        if (assemblies.Count == 1)
-        {
-            terminal.SetColor(TerminalColor.DarkGray);
-            terminal.Append(" - ");
-            terminal.ResetColor();
-            AppendAssemblyLinkTargetFrameworkAndArchitecture(terminal, assemblies[0]);
-        }
-
-        terminal.ResetColor();
-        terminal.AppendLine();
-
-        if (WasCancelled)
-        {
-            terminal.Append(TerminalResources.Aborted);
-            terminal.AppendLine();
-        }
-
-        string durationText = $"{SingleIndentation}{TerminalResources.DurationLowercase}: ";
-        TimeSpan runDuration = _testExecutionStartTime != null && _testExecutionEndTime != null ? (_testExecutionEndTime - _testExecutionStartTime).Value : TimeSpan.Zero;
-        terminal.Append(durationText);
-        AppendLongDuration(terminal, runDuration, wrapInParentheses: false, colorize: false);
-        terminal.AppendLine();
-    }
-
-    /// <summary>
-    /// Appends the opt-in "Slowest tests" section, ranking the longest-running tests by their reported execution
-    /// duration. For a single assembly a flat list is rendered; for the multi-assembly orchestrator each assembly
-    /// gets its own sub-list so the ranking stays scoped per assembly. No-op when the feature is off or when no
-    /// timed tests were recorded.
-    /// </summary>
-    private void AppendSlowestTests(ITerminal terminal, List<TestProgressState> assemblies)
-    {
-        int count = _options.SlowestTestsCount;
-        if (count <= 0)
-        {
-            return;
-        }
-
-        if (_options.ShowAssembly && assemblies.Count > 1)
-        {
-            bool headerWritten = false;
-            foreach (TestProgressState assembly in assemblies)
-            {
-                IReadOnlyList<(string DisplayName, TimeSpan Duration)> slowest = assembly.GetSlowestTests(count);
-                if (slowest.Count == 0)
-                {
-                    continue;
-                }
-
-                if (!headerWritten)
-                {
-                    terminal.AppendLine();
-                    terminal.AppendLine(TerminalResources.SlowestTests);
-                    headerWritten = true;
-                }
-
-                terminal.Append(SingleIndentation);
-                AppendAssemblyLinkTargetFrameworkAndArchitecture(terminal, assembly);
-                terminal.AppendLine();
-                foreach ((string displayName, TimeSpan duration) in slowest)
-                {
-                    terminal.Append(DoubleIndentation);
-                    AppendSlowestTestLine(terminal, displayName, duration);
-                }
-            }
-
-            return;
-        }
-
-        // Single assembly (in-process host, or the orchestrator with a single assembly): a flat list.
-        IReadOnlyList<(string DisplayName, TimeSpan Duration)> tests = assemblies.Count == 1
-            ? assemblies[0].GetSlowestTests(count)
-            : [];
-        if (tests.Count == 0)
-        {
-            return;
-        }
-
-        terminal.AppendLine();
-        terminal.AppendLine(TerminalResources.SlowestTests);
-        foreach ((string displayName, TimeSpan duration) in tests)
-        {
-            terminal.Append(SingleIndentation);
-            AppendSlowestTestLine(terminal, displayName, duration);
-        }
-    }
-
-    /// <summary>
     /// Appends the retry accounting lines that sit between the skipped count and the duration:
     /// <c>flaky: N</c> (tests that failed at least once but eventually passed) and
     /// <c>retried: N tests, M extra runs</c>. Both are omitted entirely when nothing was retried, so a run without
@@ -436,280 +245,4 @@ internal sealed partial class TerminalTestReporter
             terminal.ResetColor();
         }
     }
-
-    /// <summary>
-    /// Appends the "Flaky tests" section listing, by name, the tests that failed at least once but whose final
-    /// attempt passed. Retried tests that never recovered are deliberately not listed: they are already reported as
-    /// failures with their full error output, so a second listing would only duplicate. For a single assembly a flat
-    /// list is rendered; the multi-assembly orchestrator groups per assembly. No-op when the feature is off or when
-    /// no test was flaky.
-    /// </summary>
-    private void AppendFlakyTests(ITerminal terminal, List<TestProgressState> assemblies)
-    {
-        if (!_options.ShowFlakyTests)
-        {
-            return;
-        }
-
-        if (_options.ShowAssembly && assemblies.Count > 1)
-        {
-            bool headerWritten = false;
-            foreach (TestProgressState assembly in assemblies)
-            {
-                IReadOnlyList<(string DisplayName, int Attempts)> flaky = assembly.GetFlakyTests();
-                if (flaky.Count == 0)
-                {
-                    continue;
-                }
-
-                if (!headerWritten)
-                {
-                    terminal.AppendLine();
-                    terminal.AppendLine(TerminalResources.FlakyTests);
-                    headerWritten = true;
-                }
-
-                terminal.Append(SingleIndentation);
-                AppendAssemblyLinkTargetFrameworkAndArchitecture(terminal, assembly);
-                terminal.AppendLine();
-                foreach ((string displayName, int attempts) in flaky)
-                {
-                    terminal.Append(DoubleIndentation);
-                    AppendFlakyTestLine(terminal, displayName, attempts);
-                }
-            }
-
-            return;
-        }
-
-        IReadOnlyList<(string DisplayName, int Attempts)> tests = assemblies.Count == 1
-            ? assemblies[0].GetFlakyTests()
-            : [];
-        if (tests.Count == 0)
-        {
-            return;
-        }
-
-        terminal.AppendLine();
-        terminal.AppendLine(TerminalResources.FlakyTests);
-        foreach ((string displayName, int attempts) in tests)
-        {
-            terminal.Append(SingleIndentation);
-            AppendFlakyTestLine(terminal, displayName, attempts);
-        }
-    }
-
-    private static void AppendFlakyTestLine(ITerminal terminal, string displayName, int attempts)
-    {
-        terminal.Append(MakeControlCharactersVisible(displayName, true));
-        terminal.SetColor(TerminalColor.DarkGray);
-        terminal.Append(' ');
-        terminal.Append(TerminalResources.FlakyTransition);
-        terminal.Append(" (");
-        terminal.Append(string.Format(CultureInfo.CurrentCulture, TerminalResources.FlakyAttempts, attempts));
-        terminal.Append(')');
-        terminal.ResetColor();
-        terminal.AppendLine();
-    }
-
-    private static void AppendSlowestTestLine(ITerminal terminal, string displayName, TimeSpan duration)
-    {
-        AppendLongDuration(terminal, duration, wrapInParentheses: false);
-        terminal.Append(' ');
-        terminal.Append(MakeControlCharactersVisible(displayName, true));
-        terminal.AppendLine();
-    }
-
-    internal void AppendCoverageSummary(IReadOnlyList<CoverageScopeSummary> scopes, IReadOnlyList<TestCoverageThresholdMessage> thresholds)
-        => _terminalWithProgress.WriteToTerminal(terminal => AppendCoverageSummary(terminal, scopes, thresholds));
-
-    private static void AppendCoverageSummary(ITerminal terminal, IReadOnlyList<CoverageScopeSummary> scopes, IReadOnlyList<TestCoverageThresholdMessage> thresholds)
-    {
-        CoverageScopeSummary[] summaryScopes =
-            [.. scopes.Where(scope => scope.Scope.Level is CoverageScopeLevel.Overall or CoverageScopeLevel.Module)];
-        if (summaryScopes.Length == 0 && thresholds.Count == 0)
-        {
-            return;
-        }
-
-        terminal.AppendLine();
-
-        if (summaryScopes.Length > 0)
-        {
-            terminal.AppendLine($"{SingleIndentation}{TerminalResources.CodeCoverageSummary}");
-
-            foreach (CoverageScopeSummary scope in summaryScopes)
-            {
-                string scopeLabel = GetCoverageScopeLabel(scope.Scope);
-                Dictionary<(CoverageMetric Metric, string? CustomMetricName), int> metricCounts = [];
-                foreach (CoverageMetricResult metric in scope.Metrics)
-                {
-                    (CoverageMetric Metric, string? CustomMetricName) metricKey = (metric.Metric, metric.CustomMetricName);
-                    metricCounts.TryGetValue(metricKey, out int count);
-                    metricCounts[metricKey] = count + 1;
-                }
-
-                foreach (CoverageMetricResult metric in scope.Metrics)
-                {
-                    terminal.Append(DoubleIndentation);
-                    string coverageValue = metric.HasCoverableData
-                        ? $"{metric.Percentage.ToString("F1", CultureInfo.InvariantCulture)}%"
-                        : TerminalResources.CoverageNoData;
-                    string metricLabel = GetCoverageMetricLabel(metric.Metric, metric.CustomMetricName);
-                    if (metricCounts[(metric.Metric, metric.CustomMetricName)] > 1)
-                    {
-                        metricLabel = string.Format(
-                            CultureInfo.InvariantCulture,
-                            TerminalResources.CoverageMetricWithProducer,
-                            metricLabel,
-                            MakeControlCharactersVisible(metric.ProducerId, true));
-                    }
-
-                    terminal.AppendLine($"{scopeLabel} - {metricLabel}: {coverageValue}");
-                }
-            }
-        }
-
-        if (thresholds.Count > 0)
-        {
-            // Only separate from the coverage block above when it was actually rendered; otherwise the
-            // unconditional blank line at the top of the method already provides the single leading blank.
-            if (summaryScopes.Length > 0)
-            {
-                terminal.AppendLine();
-            }
-
-            terminal.AppendLine($"{SingleIndentation}{TerminalResources.CoverageThresholdResults}");
-
-            Dictionary<(CoverageScope Scope, CoverageMetric Metric, string? CustomMetricName, CoverageAggregation Aggregation, CoverageScopeLevel? AggregatedOver), int> thresholdCounts = [];
-            foreach (TestCoverageThresholdMessage threshold in thresholds)
-            {
-                (CoverageScope Scope, CoverageMetric Metric, string? CustomMetricName, CoverageAggregation Aggregation, CoverageScopeLevel? AggregatedOver) thresholdKey =
-                    (threshold.Scope, threshold.Metric, threshold.CustomMetricName, threshold.Aggregation, threshold.AggregatedOver);
-                thresholdCounts.TryGetValue(thresholdKey, out int count);
-                thresholdCounts[thresholdKey] = count + 1;
-            }
-
-            foreach (TestCoverageThresholdMessage threshold in thresholds)
-            {
-                bool passed = threshold.Passed;
-                terminal.SetColor(passed ? TerminalColor.DarkGreen : TerminalColor.DarkRed);
-                terminal.Append(DoubleIndentation);
-                string comparison;
-                if (threshold.HasCoverableData)
-                {
-                    (string actual, string required) = FormatThresholdPercentages(threshold);
-                    comparison = string.Format(
-                        CultureInfo.InvariantCulture,
-                        passed ? TerminalResources.CoverageThresholdPassed : TerminalResources.CoverageThresholdFailed,
-                        actual,
-                        required);
-                }
-                else
-                {
-                    comparison = passed
-                        ? TerminalResources.CoverageThresholdNoDataPassed
-                        : TerminalResources.CoverageThresholdNoDataFailed;
-                }
-
-                string thresholdLabel = GetCoverageThresholdLabel(threshold);
-                if (thresholdCounts[(threshold.Scope, threshold.Metric, threshold.CustomMetricName, threshold.Aggregation, threshold.AggregatedOver)] > 1)
-                {
-                    thresholdLabel = string.Format(
-                        CultureInfo.InvariantCulture,
-                        TerminalResources.CoverageMetricWithProducer,
-                        thresholdLabel,
-                        MakeControlCharactersVisible(threshold.ProducerId, true));
-                }
-
-                terminal.AppendLine($"{thresholdLabel}: {comparison}");
-                terminal.ResetColor();
-            }
-        }
-    }
-
-    private static (string Actual, string Required) FormatThresholdPercentages(TestCoverageThresholdMessage threshold)
-    {
-        string actual = threshold.ActualPercentage.ToString("F1", CultureInfo.InvariantCulture);
-        string required = threshold.RequiredPercentage.ToString("F1", CultureInfo.InvariantCulture);
-
-        // Rounding a failed near-boundary result to one decimal can make both operands look equal.
-        // Increase precision only for that ambiguous case so the rendered comparison remains truthful.
-        if (!threshold.Passed && string.Equals(actual, required, StringComparison.Ordinal))
-        {
-            actual = threshold.ActualPercentage.ToString("G17", CultureInfo.InvariantCulture);
-            required = threshold.RequiredPercentage.ToString("G17", CultureInfo.InvariantCulture);
-        }
-
-        return (actual, required);
-    }
-
-    // Renders the scope identity used in the coverage summary; the whole-run (Overall) scope is shown
-    // with a localized "Total" label, every other scope uses its own name.
-    private static string GetCoverageScopeLabel(CoverageScope scope)
-        => scope.Level == CoverageScopeLevel.Overall
-            ? TerminalResources.CoverageScopeOverall
-            : MakeControlCharactersVisible(scope.Name ?? GetCoverageScopeLevelLabel(scope.Level), true);
-
-    // Maps the CoverageMetric enum to a localized label so localized runs don't render the English enum
-    // identifier; falls back to the identifier (or custom name) for any member without a resource.
-    private static string GetCoverageMetricLabel(CoverageMetric metric, string? customMetricName)
-        => metric switch
-        {
-            CoverageMetric.Line => TerminalResources.CoverageMetricLine,
-            CoverageMetric.Statement => TerminalResources.CoverageMetricStatement,
-            CoverageMetric.Branch => TerminalResources.CoverageMetricBranch,
-            CoverageMetric.Method => TerminalResources.CoverageMetricMethod,
-            CoverageMetric.Function => TerminalResources.CoverageMetricFunction,
-            CoverageMetric.Block => TerminalResources.CoverageMetricBlock,
-            CoverageMetric.Instruction => TerminalResources.CoverageMetricInstruction,
-            CoverageMetric.Region => TerminalResources.CoverageMetricRegion,
-            CoverageMetric.Class => TerminalResources.CoverageMetricClass,
-            CoverageMetric.Condition => TerminalResources.CoverageMetricCondition,
-            CoverageMetric.Complexity => TerminalResources.CoverageMetricComplexity,
-            CoverageMetric.Custom => MakeControlCharactersVisible(customMetricName ?? nameof(CoverageMetric.Custom), true),
-            _ => metric.ToString(),
-        };
-
-    private static string GetCoverageThresholdLabel(TestCoverageThresholdMessage threshold)
-    {
-        string scope = GetCoverageScopeLabel(threshold.Scope);
-        string metric = GetCoverageMetricLabel(threshold.Metric, threshold.CustomMetricName);
-        if (threshold.Aggregation == CoverageAggregation.None)
-        {
-            return $"{scope} - {metric}";
-        }
-
-        string aggregation = GetCoverageAggregationLabel(threshold.Aggregation);
-        string aggregatedOver = GetCoverageScopeLevelLabel(threshold.AggregatedOver.GetValueOrDefault());
-        string aggregationLabel = string.Format(
-            CultureInfo.InvariantCulture,
-            TerminalResources.CoverageAggregationOver,
-            aggregation,
-            aggregatedOver);
-
-        return $"{scope} - {metric} ({aggregationLabel})";
-    }
-
-    private static string GetCoverageAggregationLabel(CoverageAggregation aggregation)
-        => aggregation switch
-        {
-            CoverageAggregation.Total => TerminalResources.CoverageAggregationTotal,
-            CoverageAggregation.Minimum => TerminalResources.CoverageAggregationMinimum,
-            CoverageAggregation.Average => TerminalResources.CoverageAggregationAverage,
-            CoverageAggregation.Maximum => TerminalResources.CoverageAggregationMaximum,
-            _ => aggregation.ToString(),
-        };
-
-    private static string GetCoverageScopeLevelLabel(CoverageScopeLevel level)
-        => level switch
-        {
-            CoverageScopeLevel.Overall => TerminalResources.CoverageScopeOverall,
-            CoverageScopeLevel.Module => TerminalResources.CoverageScopeModule,
-            CoverageScopeLevel.Assembly => TerminalResources.CoverageScopeAssembly,
-            CoverageScopeLevel.Namespace => TerminalResources.CoverageScopeNamespace,
-            CoverageScopeLevel.Type => TerminalResources.CoverageScopeType,
-            CoverageScopeLevel.File => TerminalResources.CoverageScopeFile,
-            _ => level.ToString(),
-        };
 }
